수정

+++ src/main/java/kr/miraesoft/miraebook/repository/PublisherRepository.java
@@ -4,6 +4,6 @@ import kr.miraesoft.miraebook.domain.Publisher;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PublisherRepository extends JpaRepository<Publisher, Long>{
+public interface PublisherRepository extends JpaRepository<Publisher, Integer>{
 	
 }
